ai-app/requirements.txt
Django>=4.2.0,<5.0.0
pymongo>=4.6.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
transformers>=4.40.0
torch>=2.3.0